Adds an example unit test that reads some information from an external resource file.
Try it out by issuing:

```bash
bazel test //lib/...
```

This change also proposes a way to work around the differences between
resource file references for 'go test' and 'bazel test'. To do this we
detect the environment variable "TEST_TMPDIR" which is set by Bazel.
When this directory is available, use the passed in dirname else, the
file will be in the same directory.