Update build_rules to latest Bazel.
The Bazel build rules have gotten a little love lately and now
correctly pass environment variables which obviates the need for the
dataDir function. Now when you refer to a file in your tests you can
refer to it the same way you would if you were using the go tooling.